added test to narrow down bug. it is not in the product variant function/ api call. it might be on the printify end at the create product level. one solution is to do another fetch variants at the product page and use that instead of the variants on the product. before doing that will see if can find out why the product has the wrong variants on it.
