+++ lib/printify/service.ts
@@ -13,6 +13,7 @@ import {
 } from "@/interfaces/PrintifyTypes";
 import { envServer } from "@/lib/env/server";
 import { getPromptFromImageId } from "@/db/image";
+import { fetchProductVariants } from "./fetchProductVariants";
 
 async function constructPrintifyProductRequest({
     printifyImageId,
@@ -216,21 +217,6 @@ export async function createPrintifyProduct({
     );
     const productData = await productResponse.json();
     return productData;
-}
-
-export async function fetchProductVariants(
-    blueprintId: number,
-    printProviderId: number,
-): Promise<Variant[]> {
-    const endpoint = `${PRINTIFY_BASE_URL}/v1/catalog/blueprints/${blueprintId}/print_providers/${printProviderId}/variants.json`;
-    const options: RequestInit = {
-        headers: {
-            "Content-Type": "application/json",
-            authorization: `Bearer ${envServer.PRINTIFY_API_TOKEN}`,
-        },
-    };
-    const response = await (await fetch(endpoint, options)).json();
-    return response.variants;
 }
 
 export function mapProductDetails(variant: Variant): EssentialProductDetails {
